Address code review feedback - improve consistency and remove debug code

+++ includes/class-social-feed-cache.php
@@ -101,9 +101,11 @@ class Social_Feed_Cache {
 		}
 
 		$files = glob( $this->cache_dir . '/*.cache' );
-		foreach ( $files as $file ) {
-			if ( is_file( $file ) ) {
-				unlink( $file );
+		if ( is_array( $files ) ) {
+			foreach ( $files as $file ) {
+				if ( is_file( $file ) ) {
+					unlink( $file );
+				}
 			}
 		}
